Refactoring Filesystem to ReportDirectory
