+++ piperider_cli/recipes/__init__.py
@@ -11,9 +11,8 @@ from ruamel import yaml
 
 import piperider_cli.dbtutil as dbtutil
 from piperider_cli import get_run_json_path, load_jinja_template, load_json
-from piperider_cli.configuration import PIPERIDER_WORKSPACE_NAME, Configuration
+from piperider_cli.configuration import Configuration, PIPERIDER_WORKSPACE_NAME
 from piperider_cli.error import RecipeConfigException
-from piperider_cli.filesystem import FileSystem
 from piperider_cli.recipes.utils import InteractiveStopException
 
 PIPERIDER_RECIPES_SCHEMA_PATH = os.path.join(os.path.dirname(__file__), 'recipe_schema.json')
@@ -153,7 +152,7 @@ class RecipeModel:
 
     def get_run_report(self):
         if not self.is_file_specified():
-            filesystem = FileSystem()
+            filesystem = Configuration.instance().activate_report_directory()
             return get_run_json_path(filesystem.get_output_dir())
         return self.file
